+++ DTCD-StyleSystem/src/base-components/BaseColorPicker/BaseColorPicker.js
@@ -1,4 +1,5 @@
 import html from './BaseColorPicker.html';
+import styles from './BaseColorPicker.scss';
 
 const colors = [
   { name: 'title', val: 'rgba(37, 34, 48, 1)' },
@@ -32,6 +33,7 @@ export default class BaseColorPicker extends HTMLElement {
   #selectedPreview;
   #colorListClickHandler;
   #selectedPreviewClickHandler;
+  #value;
 
   static get observedAttributes() {
     return ['value', 'disabled'];
@@ -46,28 +48,32 @@ export default class BaseColorPicker extends HTMLElement {
     this.attachShadow({ mode: 'open' });
     this.shadowRoot.appendChild(template.content.cloneNode(true));
 
-    this.#picker = this.shadowRoot.querySelector('.picker');
-    this.#colorList = this.shadowRoot.querySelector('.color-list');
-    this.#selectedPreview = this.shadowRoot.querySelector('#selected-color');
+    this.#picker = this.shadowRoot.querySelector('.BaseColorPicker');
+    this.#colorList = this.shadowRoot.querySelector('.ColorList');
+    this.#selectedPreview = this.shadowRoot.querySelector('.Field');
 
-    this.value = '#252230';
+    const style = document.createElement('style');
+    this.shadowRoot.appendChild(style);
+    style.appendChild(document.createTextNode(styles));
+
+    this.#value = '#252230';
 
     this.#colorListClickHandler = ({ target }) => {
-      const classes = ['color-variant', 'color-preview'];
+      const classes = ['SelectedColor', 'ColorPreview'];
 
       if (!classes.includes(target.className)) return;
 
       const color = target.getAttribute('data-color');
-      this.value = color;
+      this.#value = color;
       this.#setSelectedColorBackground(color);
       this.dispatchEvent(new Event('input', { bubbles: true }));
 
-      if (this.#colorList.classList.contains('open')) this.#toggleColorList();
+      this.toggle(false);
     };
 
     this.#selectedPreviewClickHandler = () => {
       if (!this.disabled) {
-        this.#toggleColorList();
+        this.toggle();
       }
     };
 
@@ -84,30 +90,48 @@ export default class BaseColorPicker extends HTMLElement {
     else this.removeAttribute('disabled');
   }
 
-  #setSelectedColorBackground(color = '#252230') {
-    this.#selectedPreview.querySelector('.color-preview').style.backgroundColor = color;
+  get value() {
+    return this.#value;
   }
 
-  #addColorVariant(color) {
-    const variant = document.createElement('div');
-    variant.className = 'color-variant';
-    variant.setAttribute('data-color', color);
+  set value(value) {
+    if (value) this.setAttribute('value', value);
+    else this.removeAttribute('value');
+  }
+
+  #setSelectedColorBackground(color = '#252230') {
+    this.#selectedPreview.querySelector('.ColorPreview').style.backgroundColor = color;
+  }
+
+  #addColorSelected(color) {
+    const selected = document.createElement('div');
+    selected.className = 'SelectedColor';
+    selected.setAttribute('data-color', color);
 
     const preview = document.createElement('div');
-    preview.className = 'color-preview';
+    preview.className = 'ColorPreview';
     preview.style.backgroundColor = color;
     preview.setAttribute('data-color', color);
 
-    variant.appendChild(preview);
-    this.#colorList.appendChild(variant);
+    selected.appendChild(preview);
+    this.#colorList.appendChild(selected);
   }
 
-  #toggleColorList() {
-    this.#colorList.classList.toggle('open');
+  toggle(doOpen) {
+    if (doOpen == undefined) {
+      this.#picker.classList.toggle('opened');
+      return;
+    }
+
+    if (doOpen) {
+      this.#picker.classList.add('opened');
+    } else {
+      this.#picker.classList.remove('opened');
+    }
   }
 
   connectedCallback() {
-    colors.forEach(c => this.#addColorVariant(c.val));
+    colors.forEach(c => this.#addColorSelected(c.val));
   }
 
   disconnectedCallback() {
@@ -119,12 +143,12 @@ export default class BaseColorPicker extends HTMLElement {
     if (attrName === 'disabled') {
       this.#picker.classList.toggle('disabled');
       if (this.disabled) {
-        this.#colorList.classList.remove('open');
+        this.#picker.classList.remove('opened');
       }
     }
 
     if (attrName === 'value') {
-      this.value = newValue;
+      this.#value = newValue;
       this.#setSelectedColorBackground(newValue);
     }
   }